refactor: add saperate row component with hooks for token search list, reduce number of list renders
